LayoutTests/imported/w3c/web-platform-tests/editing/whitespaces/chrome-compat/forwarddelete-img.tentative-expected.txt


a   

PASS document.execCommand("forwarddelete") when "{}<img src="${src}">&nbsp;&nbsp;&nbsp;"
PASS document.execCommand("forwarddelete") when "{}<img src="${src}">&nbsp;&nbsp;&nbsp;b"
PASS document.execCommand("forwarddelete") when "a[]<img src="${src}">&nbsp;&nbsp;&nbsp;b"
PASS document.execCommand("forwarddelete") when "a&nbsp;&nbsp;&nbsp;[]<img src="${src}">&nbsp;&nbsp;&nbsp;b"
PASS document.execCommand("forwarddelete") when "a&nbsp;&nbsp;&nbsp;[]<img src="${src}">b"
FAIL document.execCommand("forwarddelete") when "a&nbsp;&nbsp;&nbsp;[]<img src="${src}">" assert_equals: expected "a&nbsp; &nbsp;" but got "a &nbsp;&nbsp;"
FAIL document.execCommand("forwarddelete") when "a&nbsp;&nbsp; []<img src="${src}">" assert_equals: expected "a&nbsp; &nbsp;" but got "a &nbsp;&nbsp;"
